Use this code - Fixed most typos and errors
Went through and fixed most of the error messages and typo messages.  Having issues fixing errors within the sounds() loop however.  Error message of UselessBox:254:26: error: lvalue required as left operand of assignment
     if (pinstate==1 && i=1){ 
is showing up.  It is also having issues with the lvalue on the rest of the else if statements.  I tried adding the pinstate as an int in the else if statements and at the top with the rest of the declared variables but neither is working.  Dont know what to try now.