Use this code - Added repeat function
Created a first draft of the repeat function for the box to "get mad" if the same switch is pressed over and over.  This function is currently commented out as it is not necessary for the operation of the code and could complicate it.  It is also not complete and doesn't have a place in either of the loops yet.  Again, just a first draft of the function