Use this code
Can recognize when switch is pressed, but only once for some reason.  May work when whole circuit is put together.  When loop in line 417 is active, it will run continuously and wont exit the loop.  distancemeasure() will continuously run.  Maybe use an interrupt in a different way to activate something when the person is close or when they walk away