Stepper, Servo, Switches and Speaker Isolation
This code is functional for the stepper motor movement between all 5 switches, the servo activation at each switch, the speaker tones at each switch. It works for multiple switches and hits them in an array looping 1-5. I.E., if switches 1 and  5 are pressed, it will loop back and forth between 1 and 5, trying to hit each one.